+++ processor/src/test/resources/expected/Arez_EmptyModel.java
@@ -1,5 +1,5 @@
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -14,7 +14,7 @@ public final class Arez_EmptyModel extends EmptyModel implements Disposable {
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;