+++ processor/src/test/resources/expected/Arez_OverrideNamesInModel.java
@@ -2,6 +2,7 @@ import java.text.ParseException;
 import java.util.Objects;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -20,7 +21,7 @@ public final class Arez_OverrideNamesInModel extends OverrideNamesInModel implem
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;