+++ processor/src/test/resources/expected/NestedNestedModel_Something_Arez_BasicActionModel.java
@@ -1,5 +1,5 @@
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -15,7 +15,7 @@ public final class NestedNestedModel_Something_Arez_BasicActionModel extends Nes
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;