+++ processor/src/test/resources/expected/com/example/component_id/Arez_ObjectComponentId.java
@@ -1,7 +1,7 @@
 package com.example.component_id;
 
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -12,7 +12,7 @@ import org.realityforge.arez.Observable;
 public final class Arez_ObjectComponentId extends ObjectComponentId implements Disposable {
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;