+++ processor/src/test/resources/expected/com/example/computed/Arez_ComputedWithNameVariationsModel.java
@@ -3,6 +3,7 @@ package com.example.computed;
 import java.util.Objects;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -20,7 +21,7 @@ public final class Arez_ComputedWithNameVariationsModel extends ComputedWithName
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;