+++ processor/src/test/resources/expected/com/example/context_ref/Arez_AnnotatedComponent.java
@@ -2,6 +2,7 @@ package com.example.context_ref;
 
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -16,7 +17,7 @@ final class Arez_AnnotatedComponent extends AnnotatedComponent implements Dispos
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;