+++ processor/src/test/resources/expected/com/example/inheritance/Arez_MyModel.java
@@ -1,7 +1,7 @@
 package com.example.inheritance;
 
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -17,7 +17,7 @@ final class Arez_MyModel extends MyModel implements Disposable {
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;