+++ processor/src/test/resources/expected/com/example/observable_ref/Arez_DefaultRefNameModel.java
@@ -2,6 +2,7 @@ package com.example.observable_ref;
 
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -17,7 +18,7 @@ public final class Arez_DefaultRefNameModel extends DefaultRefNameModel implemen
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;