+++ processor/src/test/resources/expected/com/example/override_generics/Arez_GenericsModel.java
@@ -1,7 +1,7 @@
 package com.example.override_generics;
 
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -17,7 +17,7 @@ public final class Arez_GenericsModel extends GenericsModel implements Disposabl
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;