+++ processor/src/test/resources/expected/com/example/repository/Arez_CompleteRepositoryExampleRepository.java
@@ -3,6 +3,7 @@ package com.example.repository;
 import java.util.Collection;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -18,7 +19,7 @@ public final class Arez_CompleteRepositoryExampleRepository extends CompleteRepo
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;