+++ processor/src/test/resources/expected/com/example/repository/NestedModel_Arez_BasicActionModel.java
@@ -1,7 +1,7 @@
 package com.example.repository;
 
 import javax.annotation.Generated;
-import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -19,7 +19,7 @@ public final class NestedModel_Arez_BasicActionModel extends NestedModel.BasicAc
 
   private OnDispose $$arez$$_onDispose;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;