+++ processor/src/test/resources/expected/com/example/tracked/Arez_BasicTrackedWithExceptionsModel.java
@@ -3,6 +3,7 @@ package com.example.tracked;
 import java.text.ParseException;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -19,7 +20,7 @@ public final class Arez_BasicTrackedWithExceptionsModel extends BasicTrackedWith
 
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;