+++ processor/src/test/resources/expected/com/example/type_params/Arez_ConcreteModel.java
@@ -5,6 +5,7 @@ import java.io.Writer;
 import java.util.Objects;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
 import org.realityforge.arez.ArezContext;
 import org.realityforge.arez.Component;
@@ -18,7 +19,7 @@ import org.realityforge.braincheck.Guards;
 public final class Arez_ConcreteModel<W extends Writer> extends ConcreteModel<W> implements Disposable {
   private boolean $$arez$$_disposed;
 
-  @Nonnull
+  @Nullable
   private final ArezContext $$arez$$_context;
 
   private final Component $$arez$$_component;